# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhoneVerify API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserInfo {
  id: number;
  name: string;
  email: string;
}

export interface AuthResponse {
  token: string;
  user: UserInfo;
}

export interface AddPhoneRequest {
  imei: string;
  imageBase64: string;
  brand?: string;
  model?: string;
}

export interface Layer1Result {
  extractedImei: string;
  matchPercentage: number;
  verified: boolean;
  message: string;
}

export interface TrustScoreBreakdown {
  imeiPhotoMatch: number;
  tacCoherence: number;
  imageQuality: number;
  registrationMetadata: number;
  geographicBaseline: number;
}

export interface Layer2Result {
  score: number;
  riskLevel: string;
  breakdown: TrustScoreBreakdown;
}

export interface Layer3Result {
  forgeryDetected: boolean;
  fontInconsistency: boolean;
  colorAnomaly: boolean;
  edgeVariance: boolean;
  barcodeAlignment: boolean;
  message: string;
}

export interface PhoneAnalysisResult {
  success: boolean;
  imei: string;
  layer1: Layer1Result;
  layer2: Layer2Result;
  layer3: Layer3Result;
  trustScore: number;
  finalVerdict: string;
  message: string;
  savedToDatabase: boolean;
}

export interface SearchRequest {
  imei: string;
}

export interface StoredPhoneData {
  id: number;
  imei: string;
  brand?: string;
  model?: string;
  trustScore: number;
  layer1Score: number;
  layer2Score: number;
  layer3Result: string;
  finalVerdict: string;
  verificationCount: number;
  createdAt: string;
}

export interface SearchResult {
  found: boolean;
  imei: string;
  isSecure: boolean;
  message: string;
  searchCount: number;
  phone?: StoredPhoneData;
}

export interface ErrorResponse {
  error: string;
  message: string;
}