# lib/api-zod/src/generated/types/addPhoneRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhoneVerify API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddPhoneRequest {
  imei: string;
  imageBase64: string;
  brand?: string;
  model?: string;
}