# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhoneVerify API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserInfo } from "./userInfo";

export interface AuthResponse {
  token: string;
  user: UserInfo;
}