# lib/api-zod/src/generated/types/layer1Result.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhoneVerify API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Layer1Result {
  extractedImei: string;
  matchPercentage: number;
  verified: boolean;
  message: string;
}