# lib/api-zod/src/generated/types/layer2Result.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhoneVerify API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrustScoreBreakdown } from "./trustScoreBreakdown";

export interface Layer2Result {
  score: number;
  riskLevel: string;
  breakdown: TrustScoreBreakdown;
}