# lib/api-zod/src/generated/types/layer3Result.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhoneVerify API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Layer3Result {
  forgeryDetected: boolean;
  fontInconsistency: boolean;
  colorAnomaly: boolean;
  edgeVariance: boolean;
  barcodeAlignment: boolean;
  message: string;
}