# lib/api-zod/src/generated/types/phoneAnalysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhoneVerify API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Layer1Result } from "./layer1Result";
import type { Layer2Result } from "./layer2Result";
import type { Layer3Result } from "./layer3Result";

export interface PhoneAnalysisResult {
  success: boolean;
  imei: string;
  layer1: Layer1Result;
  layer2: Layer2Result;
  layer3: Layer3Result;
  trustScore: number;
  finalVerdict: string;
  message: string;
  savedToDatabase: boolean;
}