# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhoneVerify API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoredPhoneData } from "./storedPhoneData";

export interface SearchResult {
  found: boolean;
  imei: string;
  isSecure: boolean;
  message: string;
  searchCount: number;
  phone?: StoredPhoneData;
}