# lib/api-zod/src/generated/types/storedPhoneData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhoneVerify API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StoredPhoneData {
  id: number;
  imei: string;
  brand?: string;
  model?: string;
  trustScore: number;
  layer1Score: number;
  layer2Score: number;
  layer3Result: string;
  finalVerdict: string;
  verificationCount: number;
  createdAt: string;
}