# lib/api-zod/src/generated/types/trustScoreBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PhoneVerify API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrustScoreBreakdown {
  imeiPhotoMatch: number;
  tacCoherence: number;
  imageQuality: number;
  registrationMetadata: number;
  geographicBaseline: number;
}